About activivty
Created an about activity. Styled both the main and activity layout files. Main got a header and a button to get to the about activity. In the about activity a header was also added. One constrainlayout were created to handle the description text that were created by three different textviews. To style the description a drawable border were created and last a button to be able to navigate back to Main activity.